+++ src/test/kotlin/com/wutsi/codegen/kotlin/sdk/SdkPomCodeGeneratorTest.kt
@@ -33,7 +33,7 @@ internal class SdkPomCodeGeneratorTest {
         assertTrue(file.exists())
 
         val result = file.readText()
-        val expected = IOUtils.toString(SdkPomCodeGenerator::class.java.getResourceAsStream("/sdk/kotlin/pom.xml"))
+        val expected = IOUtils.toString(SdkPomCodeGenerator::class.java.getResourceAsStream("/kotlin/sdk/pom.xml"))
         assertEquals(expected.trimIndent(), result.trimIndent())
     }
 }